instalily-case-backend/pyproject.toml
[project]
name = "instalily-case-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "openai>=2.21.0",
]

[project.scripts]
start = "app.main:run_server"

[tool.uv]
package = true
